+++ src/promptflow/CHANGELOG.md
@@ -1,7 +1,10 @@
 # Release History
 
 ## 1.1.0 (Upcoming)
+
 ### Features Added
+- Add `pfazure flow list` to list flows from Azure AI.
+- Display node status in run visualize page graph view.
 
 ### Bugs Fixed